made load partitions multithreaded
