+++ src/persist_operations/data_initializer/load_table/load_table_files.rs
@@ -2,14 +2,10 @@ use std::{sync::Arc, time::Duration};
 
 use crate::{
     app::AppContext,
-    persist_io::TableFile,
-    persist_operations::{
-        data_initializer::load_tasks::{PartitionToLoad, TableToLoad},
-        serializers::{self, TableMetadataFileContract},
-    },
+    persist_operations::data_initializer::load_tasks::{PartitionToLoad, TableToLoad},
 };
 
-use super::{LoadedTable, LoadedTableItem};
+use super::LoadedTable;
 
 pub async fn load_table_files(
     app: Arc<AppContext>,
@@ -17,11 +13,33 @@ pub async fn load_table_files(
 ) -> Arc<LoadedTable> {
     let loaded_table = Arc::new(LoadedTable::new());
 
+    let mut to_load = Vec::with_capacity(app.settings.init_partitions_threads_amount);
+
+    for _ in 0..app.settings.init_partitions_threads_amount {
+        to_load.push(tokio::spawn(load_table_files_loop(
+            app.clone(),
+            table_to_load.clone(),
+            loaded_table.clone(),
+        )));
+    }
+
+    for handle in to_load {
+        handle.await.unwrap();
+    }
+
+    loaded_table
+}
+
+async fn load_table_files_loop(
+    app: Arc<AppContext>,
+    table_to_load: Arc<TableToLoad>,
+    loaded_table: Arc<LoadedTable>,
+) {
     loop {
         match table_to_load.get_next().await {
             Some(next_task) => match next_task {
                 PartitionToLoad::Load(file_name) => {
-                    load_table_file(
+                    super::load_table_file(
                         table_to_load.table_name.as_str(),
                         file_name,
                         &loaded_table,
@@ -36,82 +54,6 @@ pub async fn load_table_files(
             None => {
                 tokio::time::sleep(Duration::from_millis(100)).await;
             }
-        }
-    }
-
-    loaded_table
-}
-
-async fn load_table_file(
-    table_name: &str,
-    file_name: String,
-    loaded_table: &Arc<LoadedTable>,
-    app: &Arc<AppContext>,
-) {
-    let table_file = TableFile::from_file_name(file_name.as_str());
-
-    if let Err(err) = table_file {
-        app.logs.add_error(
-            Some(file_name.to_string()),
-            crate::app::logs::SystemProcess::Init,
-            "init_tables".to_string(),
-            format!("Error loading table file {}: {}", file_name, err),
-            None,
-        );
-        return;
-    }
-
-    let table_file = table_file.unwrap();
-
-    let content = app
-        .persist_io
-        .load_table_file(table_name, &table_file)
-        .await;
-
-    if let Some(content) = content.as_ref() {
-        match get_item(&table_file, content) {
-            Ok(item) => {
-                loaded_table.add(item).await;
-                app.init_state
-                    .update_file_is_loaded(file_name.as_str())
-                    .await;
-            }
-            Err(err) => {
-                if app.settings.skip_broken_partitions {
-                    println!(
-                        "Skipping file {}. Reason: {:?}",
-                        table_file.get_file_name().as_str(),
-                        err
-                    );
-                } else {
-                    panic!(
-                            "Partition is broken. Stopping initialization because of the file {}/{}. Reason: {:?}",
-                            table_name,
-                            table_file.get_file_name().as_str(),
-                            err
-                        );
-                }
-            }
-        }
-    }
-}
-
-fn get_item(table_file: &TableFile, content: &[u8]) -> Result<LoadedTableItem, String> {
-    match table_file {
-        TableFile::TableAttributes => {
-            let table_metadata = TableMetadataFileContract::parse(content);
-            let result = LoadedTableItem::TableAttributes(table_metadata.into());
-            return Ok(result);
-        }
-        TableFile::DbPartition(partition_key) => {
-            let db_partition = serializers::db_partition::deserialize(content)?;
-
-            let result = LoadedTableItem::DbPartition {
-                partition_key: partition_key.to_string(),
-                db_partition,
-            };
-
-            return Ok(result);
         }
     }
 }
