+++ src/persist_operations/data_initializer/load_table/mod.rs
@@ -1,8 +1,10 @@
 mod load_table;
+mod load_table_file;
 mod load_table_files;
 mod loaded_table;
 mod loaded_table_item;
 pub use load_table::load_table;
+pub use load_table_file::load_table_file;
 pub use load_table_files::load_table_files;
 pub use loaded_table::LoadedTable;
 pub use loaded_table_item::LoadedTableItem;
