+++ wwwroot/js/app.js
@@ -6,7 +6,7 @@ eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a
 
 eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--)d[e(c)]=k[c]||e(c);k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--)if(k[c])p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c]);return p}('1d m=(q(){q m(){}m.1c=q(){C\'<3 6="1b-1a">\'+\'<M><N>\'+\'<1 5="19-18: 17">O: <b 6="l" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>16: <b 6="d" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>15: <b 6="L-K" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>14: <b 6="w" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>13: <b 6="H-y" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>12: <b 6="G-y" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>11 o: <b 6="F-o" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>10 o: <b 6="E-o-D" 5="j-i: 0 0 k h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'<1>O Z B A: <b 6="B-A" 5="j-i: 0 0 Y h;"></b></1>\'+\'<1><3 g="f-e"></3></1>\'+\'</N></M></3>\'};m.X=q(4){c(2.x!=4.x){2.x=4.x;a.9(\'L-K\').8=2.x.r()}c(!2.l){2.l=W;a.9(\'l\').8=\'<7 5="n: z">V</7>\'}c(2.d!=4.d.6){a.9(\'d\').8=4.d.6;2.d=4.d.6}c(2.w!=4.d.J){2.w=4.d.J;a.9(\'w\').8=2.w?\'<7 5="n: z">U</7>\':\'<7 5="n: I">T</7>\'}c(2.p!=4.p){2.p=4.p;a.9(\'B-A\').8=2.p?\'<7 5="n: z">\'+2.p+\'</7>\':\'<7 5="n: I">---</7>\'}c(2.v!=4.v){2.v=4.v;a.9(\'H-y\').8=2.v.r()}c(2.u!=4.u){2.u=4.u;a.9(\'G-y\').8=2.u.r()}c(2.t!=4.t){2.t=4.t;a.9(\'F-o\').8=2.t.r()}c(2.s!=4.s){2.s=4.s;a.9(\'E-o-D\').8=2.s.r()}};m.S=q(){c(2.l){2.l=R;a.9(\'l\').8=\'<7 5="n: Q">P</7>\'}};C m}());',62,76,'|td|this|div|data|style|id|span|innerHTML|getElementById|document||if|location|separator|statusbar|class|white|shadow|text|2px|connected|HtmlStatusBar|color|queue|masterNode|function|toString|syncQueueSize|persistAmount|httpConnections|tcpConnections|compression|tablesAmount|connections|green|node|master|return|size|sync|persist|http|tcp|gray|compress|amount|tables|table|tr|Connected|offline|red|false|updateOffline|disabled|enabled|yes|true|updateStatusbar|1px|to|Sync|Persist|Http|Tcp|Compression|Tables|Location|5px|left|padding|bar|status|layout|var'.split('|'),0,{}))
 
-eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--)d[e(c)]=k[c]||e(c);k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--)if(k[c])p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c]);return p}('5 8=(9(){9 8(){}8.A=9(){3\'<J 1f="1e"></J>\'+1d.A()+1c.A()};8.1b=9(t){5 e=\'<I>1a 19 18 17: \'+t.16+\'</I><H>15 14 13 12: \'+4.o(t.11)+\'</H>\'+\'<j Z="j j-Y j-X"><r><7>W</7><7>G i</7><7>G V</7><7>F U</7><7>F T</7></r>\';S(5 s=0,z=t.d.R(9(a,b){3 a.y>b.y?1:-1});s<z.Q;s++){5 g=z[s];e+=\'<r><6 P="O:N%">\'+g.y+\'</6><6>\'+g.i+\'</6><6>\'+g.x+"</6><6>"+4.o(g.D)+"</6><6>"+4.E(g)+"</6></r>"}3 e+"</j>"};8.E=9(d){h(d.x==0||d.i==0){3"M"}5 C=d.D/d.i;5 q=(d.x-d.i)*C;q=4.n(q);3 4.o(q)};8.k=9(f){h(f<10){3"0"+f.w(0)}3 f.w(0)};8.n=9(f){5 e=f.w(2);5 p=e.B(\'.\');h(p<0){p=e.B(\',\')}3 L(e.K(0,p))};8.o=9(c){5 m=0;h(c>=v){m=4.n(c/v);c-=m*v}5 l=0;h(c>=u){l=4.n(c/u);c-=l*u}3 4.k(m)+":"+4.k(l)+":"+4.k(c)};3 8}());',62,78,'|||return|this|var|td|th|HtmlMain|function|||seconds|progress|result|value|itm|if|loaded|table|formatSecMin|mins|hours|trunc|formatSeconds|pos|remains|tr|_i|model|60|3600|toFixed|partitions|tableName|_a|layout|indexOf|pieceDuration|secondsGone|getInitRemains|Time|Partitions|h2|h1|div|substring|parseInt|Unknown|50|width|style|length|sort|for|estimation|gone|total|TableName|bordered|striped|class||initializingSeconds|is|time|loading|Total|tablesRemains|load|to|tables|Remains|generateInit|HtmlDialog|HtmlStatusBar|main|id'.split('|'),0,{}))
+eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--)d[e(c)]=k[c]||e(c);k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--)if(k[c])p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c]);return p}('5 8=(9(){9 8(){}8.A=9(){3\'<J 1f="1e"></J>\'+1d.A()+1c.A()};8.1b=9(t){5 e=\'<I>1a 19 18 17: \'+t.16+\'</I><H>15 14 13 12: \'+4.o(t.11)+\'</H>\'+\'<j Z="j j-Y j-X"><r><7>W</7><7>G i</7><7>G V</7><7>F U</7><7>F T</7></r>\';S(5 s=0,z=t.d.R(9(a,b){3 a.y>b.y?1:-1});s<z.Q;s++){5 g=z[s];e+=\'<r><6 P="O:N%">\'+g.y+\'</6><6>\'+g.i+\'</6><6>\'+g.x+"</6><6>"+4.o(g.D)+"</6><6>"+4.E(g)+"</6></r>"}3 e+"</j>"};8.E=9(d){h(d.x==0||d.i==0){3"M"}5 C=d.D/d.i;5 q=(d.x-d.i)*C;q=4.n(q);3 4.o(q)};8.k=9(f){h(f<10){3"0"+f.w(0)}3 f.w(0)};8.n=9(f){5 e=f.w(2);5 p=e.B(\'.\');h(p<0){p=e.B(\',\')}3 L(e.K(0,p))};8.o=9(c){5 m=0;h(c>=v){m=4.n(c/v);c-=m*v}5 l=0;h(c>=u){l=4.n(c/u);c-=l*u}3 4.k(m)+":"+4.k(l)+":"+4.k(c)};3 8}());',62,78,'|||return|this|var|td|th|HtmlMain|function|||seconds|progress|result|value|itm|if|loaded|table|formatSecMin|mins|hours|trunc|formatSeconds|pos|remains|tr|_i|model|60|3600|toFixed|toLoad|tableName|_a|layout|indexOf|pieceDuration|secondsGone|getInitRemains|Time|Partitions|h2|h1|div|substring|parseInt|Unknown|50|width|style|length|sort|for|estimation|gone|total|TableName|bordered|striped|class||initializingSeconds|is|time|loading|Total|tablesRemains|load|to|tables|Remains|generateInit|HtmlDialog|HtmlStatusBar|main|id'.split('|'),0,{}))
 
 eval(function(p,a,c,k,e,d){e=function(c){return(c<a?'':e(parseInt(c/a)))+((c=c%a)>35?String.fromCharCode(c+29):c.toString(36))};if(!''.replace(/^/,String)){while(c--)d[e(c)]=k[c]||e(c);k=[function(e){return d[e]}];e=function(){return'\\w+'};c=1};while(c--)if(k[c])p=p.replace(new RegExp('\\b'+e(c)+'\\b','g'),k[c]);return p}('5 g=(c(){c g(){}g.1Z=c(f){e\'<k>U 1Y</k>\'+q.V(f.1X)+\'<k>1W</k>\'+q.17(f.1V)+\'<k>1U</k>\'+q.16(f.p)};g.17=c(f){5 9="<4 r=\\"4 4-y\\"><8><3>1T</3><3>1S</3><3>1R</3><3>p</3><3></3></8>";m(5 6=0,N=f;6<N.l;6++){5 7=N[6];9+=\'<8><2>\'+7.1Q+\'</2><2>\'+q.T(7.j)+\'</2><2>\'+7.1P+\'</2><2>\'+q.Q(7.p)+\'</2>\'+\'<2 h="10-1O: 1N">\'+\'<d><b>C:</b>\'+7.1M+\'</d>\'+\'<d><b>L:</b>\'+7.1L+\'</d>\'+\'</2></8>\'}9+=\'</4>\';e 9};g.16=c(p){5 9="<4 r=\\"4 4-y\\"><8><3>1K</3><3>1J</3><3>Y</3><3>X</3><3>B</3><3>W B</3><3>1I 1H</3></8>";5 E=0;5 D=0;5 A=0;5 z=0;m(5 6=0,M=p.1G(c(a,b){e a.j>b.j?1:-1});6<M.l;6++){5 4=M[6];5 h=\' h="n:1F" \';H(4.s<4.t){h=\' h="n:1E" \'}5 t=K J(4.t/I);5 s=K J(4.s/I);5 o="---";H(4.o){5 15=K J(4.o/I);o=15.F()}5 G="";H(!4.1D){G=\' h="Z-n: #1C" \'}9+=\'<8 \'+G+\'><2>\'+4.j+\'</2><2>\'+4.1B+\'</2><2>\'+4.14+\'</2><2>\'+4.13+\'</2><2>\'+4.12+\'</2><2>\'+4.11+\'</2>\'+\'<2\'+h+\'><d>1A: \'+t.F()+\'</d><d>1z: \'+s.F()+\'</d>\'+\'<d>1y: \'+o+\'</d>\'+1x.1w(4.1v,c(v){e 1u.1t(v)},c(v){e v},c(v){e 1s})+\'</2></8>\';E+=4.14;D+=4.13;A+=4.12;z+=4.11}9+=\'<8 h="10-1r: 1q; Z-n:1p; n:1o;"><2>1n</2><2></2><2>Y: \'+E+\'</2><2>X: \'+D+\'</2><2>B: \'+A+\'</2><2>W 1m: \'+z+\'</2>\'+\'<2></2></8>\';9+=\'</4>\';e 9};g.V=c(f){5 9="<4 r=\\"4 4-y\\"><8><3>1l</3><3>U</3><3>1k</3><3>1j</3><3>1i</3></8>";m(5 6=0,x=f;6<x.l;6++){5 7=x[6];9+=\'<8><2>\'+7.1h+\'</2><2>\'+7.1g+\'</2><2>\'+7.1f+\'</2><2>\'+7.1e+\'</2><2>\'+7.1d+\'</2></8>\'}9+=\'</4>\';e 9};g.T=c(j){5 S=j.1c(\';\');5 i="";m(5 6=0,w=S;6<w.l;6++){5 R=w[6];i+="<d>"+R+"</d>"}e i};g.Q=c(f){5 i="";m(5 6=0,u=f;6<u.l;6++){5 7=u[6];i+=\'<O r="P P-1b" h="1a-19: 18">\'+7+\'</O>\'}e i};e g}());',62,124,'||td|th|table|var|_i|itm|tr|html|||function|div|return|data|HtmlSubscribersGenerator|style|result|name|h3|length|for|color|nextPersistTime|tables|this|class|lastPersistTime|lastUpdateTime|data_3||lines_1|data_2|striped|total_indexed_records|total_records|Records||total_partitions|total_size|toISOString|lineColor|if|1000|Date|new||_a|data_1|span|badge|renderTables|line|lines|renderName|Connected|generateNodesHtml|Indexed|Partitions|DataSize|background|font|expirationIndex|recordsAmount|partitionsCount|dataSize|as_time|generateTablesHtml|generateReadersHtml|5px|left|margin|info|split|latency|compress|lastAccessed|connected|location|Latency|Compress|LastAccess|Location|records|Total|white|black|bold|weight|false|format_duration|Utils|lastPersistDuration|renderGraph|HtmlGraph|NextPersist|PersistTime|UpdateTime|persistAmount|8bc34a4f|hasCommonThread|red|green|sort|update|Last|Persist|Table|lastIncomingTime|connectedTime|10px|size|ip|id|Ip|Client|Id|Tables|readers|Readers|nodes|Nodes|generateHtml'.split('|'),0,{}))
